requirements.txt
Django==2.2
pytz==2020.1
sqlparse==0.4.1
dj-database-url==0.5.0
django-bootstrap4==2.3.1
django-tinymce==3.1.0
django-model-utils==4.0.0
gunicorn==20.0.4
Pillow==8.0.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
python-decouple==3.3
python-slugify==4.0.1
requests==2.24.0
urllib3==1.25.11
websockets==8.1
whitenoise==3.3.1